+++ index.html
@@ -13,7 +13,7 @@
 
     <link rel="manifest" href="/manifest.json" />
 
-    <title>App TP 3/AT</title>
+    <title>App TP 2</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
