--- public/locales/en.json
@@ -1,58 +1,0 @@
-{
-  "welcome": "Welcome",
-  "login": "Login",
-  "register": "Create account",
-  "sign-in": "Sign In",
-  "sign-up": "Sign Up",
-  "email": "E-mail",
-  "password": "Password",
-  "confirm-password": "Confirm password",
-  "message-email-required": "Email is required.",
-  "message-email-invalid-format": "Invalid email format.",
-  "message-password-required": "Password is required.",
-  "message-confirm-password-required": "Confirm password is required.",
-  "message-password-not-match": "Passwords do not match.",
-  "invalid-credentials": "Invalid credential.",
-  "user-creation-success": "User created successfully!!!",
-  "english": "English",
-  "spanish": "Spanish",
-  "portugues": "Portuguese",
-  "settings": "Settings",
-  "dashboard": "Dashboard",
-  "sleep": "Sleep",
-  "diaper": "Diaper",
-  "eat": "Breastfeeding",
-  "save": "Save",
-  "birth": "Birth",
-  "name": "Name",
-  "baby-data-title": "Baby data",
-  "observation": "Observation",
-  "data-hour": "Date and Time",
-  "diaper-wet": "Wet",
-  "diaper-dirty": "Dirty",
-  "diaper-both": "Both",
-  "diaper-clean": "Clean",
-  "eat-bottle": "Bottle",
-  "eat-bosom": "Breast",
-  "quantity": "Quantity",
-  "left": "Left",
-  "right": "Right",
-  "both": "Both",
-  "type": "Type",
-  "side": "Side",
-  "data-hour-start": "Start Date and Time",
-  "data-hour-end": "End Date and Time",
-  "form-updated": "Updated successfully!",
-  "form-saved": "Saved successfully!",
-  "weight": "Weight",
-  "height": "Height",
-  "field-side-required": "The breast side field is required.",
-  "field-date-required": "Date field is required.",
-  "field-start-date-required": "The start date field is required.",
-  "field-end-date-required": "The end date field is required.",
-  "field-type-required": "The type field is required.",
-  "min": "Min.",
-  "from": "From",
-  "to": "To",
-  "days": "Day(s)"
-}
