--- public/locales/es.json
@@ -1,58 +1,0 @@
-{
-  "welcome": "Bienvenido",
-  "login": "Iniciar sesión",
-  "register": "Crear una cuenta",
-  "sign-in": "Ingressar",
-  "sign-up": "Inscribirse",
-  "email": "E-mail",
-  "password": "Confirmar Clave",
-  "confirm-password": "Confirmar senha",
-  "message-email-required": "E-mail obligatorio.",
-  "message-email-invalid-format": "Formato de email no válido.",
-  "message-password-required": "Se requiere contraseña.",
-  "message-confirm-password-required": "Se requiere confirmación de contraseña.",
-  "message-password-not-match": "Las claves no coinciden.",
-  "invalid-credentials": "Credencial inválida.",
-  "user-creation-success": "Usuario creado exitosamente!!!",
-  "english": "Inglés",
-  "spanish": "Español",
-  "portugues": "Portugués",
-  "settings": "Ajustes",
-  "dashboard": "Panel",
-  "sleep": "Dormir",
-  "diaper": "Pañal",
-  "eat": "Amamantamiento",
-  "save": "Guardar",
-  "birth": "Nacimiento",
-  "name": "Nombre",
-  "baby-data-title": "Detalles del bebe",
-  "observation": "Observación",
-  "data-hour": "Fecha y hora",
-  "diaper-wet": "Húmedo",
-  "diaper-dirty": "Sucio",
-  "diaper-both": "Ambos",
-  "diaper-clean": "Limpio",
-  "eat-bottle": "Biberón",
-  "eat-bosom": "Pechos",
-  "quantity": "Cantidad",
-  "left": "Izquierdo",
-  "right": "Derecho",
-  "both": "Ambos",
-  "type": "Tipo",
-  "side": "Lado",
-  "data-hour-start": "Fecha y hora de inicio",
-  "data-hour-end": "Fecha y hora de fin",
-  "form-updated": "¡Actualizado con éxito!",
-  "form-saved": "¡Guardado exitosamente!",
-  "weight": "Peso",
-  "height": "Altura",
-  "field-side-required": "El campo del lado del pecho es obligatorio.",
-  "field-date-required": "El campo de fecha es obligatorio.",
-  "field-start-date-required": "El campo de fecha de inicio es obligatorio.",
-  "field-end-date-required": "El campo de fecha de finalización es obligatorio.",
-  "field-type-required": "El campo tipo es obligatorio.",
-  "min": "Min.",
-  "from": "De",
-  "to": "Hasta",
-  "days": "Día(s)"
-}
