--- public/locales/pt.json
@@ -1,58 +1,0 @@
-{
-  "welcome": "Seja Bem-vindo",
-  "login": "Entrar",
-  "register": "Criar conta",
-  "sign-in": "Logar",
-  "sign-up": "Registrar-se",
-  "email": "E-mail",
-  "password": "Senha",
-  "confirm-password": "Confirmar senha",
-  "message-email-required": "E-mail obrigatório.",
-  "message-email-invalid-format": "Formato de e-mail inválido.",
-  "message-password-required": "Senha obrigatória.",
-  "message-confirm-password-required": "Confirmação de senha obrigatória.",
-  "message-password-not-match": "As senhas não coincidem.",
-  "invalid-credentials": "Credenciais inválidas.",
-  "user-creation-success": "Usuário criado com sucesso!!!",
-  "english": "Inglês",
-  "spanish": "Espanhol",
-  "portugues": "Português",
-  "settings": "Configurações",
-  "dashboard": "Painel",
-  "sleep": "Sono",
-  "diaper": "Fralda",
-  "eat": "Amamentação",
-  "save": "Salvar",
-  "birth": "Nascimento",
-  "name": "Nome",
-  "baby-data-title": "Dados do bebê",
-  "observation": "Observação",
-  "data-hour": "Data e Hora",
-  "diaper-wet": "Molhada",
-  "diaper-dirty": "Suja",
-  "diaper-both": "Os dois",
-  "diaper-clean": "Limpa",
-  "eat-bottle": "Mamadeira",
-  "eat-bosom": "Seios",
-  "quantity": "Quantidade",
-  "left": "Esquerdo",
-  "right": "Direito",
-  "both": "Ambos",
-  "type": "Tipo",
-  "side": "Lado",
-  "data-hour-start": "Data e Hora inicial",
-  "data-hour-end": "Data e Hora final",
-  "form-updated": "Atualizado com sucesso!",
-  "form-saved": "Salvo com sucesso!",
-  "weight": "Peso",
-  "height": "Altura",
-  "field-side-required": "O campo lado do seio é obrigatório.",
-  "field-date-required": "O campo de data é obrigatório.",
-  "field-start-date-required": "O campo de data inicial é obrigatório.",
-  "field-end-date-required": "O campo de data final é obrigatório.",
-  "field-type-required": "O campo de tipo é obrigatório.",
-  "min": "Min.",
-  "from": "De",
-  "to": "Até",
-  "days": "Dia(s)"
-}
