--- rascunho.txt
@@ -1,5 +1,0 @@
-// ????????? Create a single supabase client for interacting with your database
-// const supabase = createClient(
-//   import.meta.env.VITE_APP_SUPABASE_URL,
-//   import.meta.env.VITE_APP_SUPABASE_API_KEY
-// );
