--- src/components/MyBox/index.jsx
@@ -1,7 +1,0 @@
-import { Box } from "@mui/material";
-
-const MyBox = ({ children, ...props }) => {
-  return <Box {...props}>{children}</Box>;
-};
-
-export default MyBox;
