+++ src/routes/index.jsx
@@ -8,7 +8,6 @@ import {
 import Dashboard from "../views/Dashboard";
 import Forms from "../views/Forms";
 import Home from "../views/Home";
-import Hometp2 from "../views/Hometeste";
 import Settings from "../views/Settings";
 import Signin from "../views/Signin";
 import Signup from "../views/Signup";
@@ -17,7 +16,6 @@ import {
   isAuthenticated,
   handleVerificationProtected,
 } from "../services/authentication";
-
 import Protected from "./protected";
 
 const router = createBrowserRouter(
@@ -26,11 +24,6 @@ const router = createBrowserRouter(
       <Route path="/" element={<Protected />}>
         <Route
           index
-          element={<Home />}
-          loader={() => handleVerificationProtected()}
-        />
-        <Route
-          path="home" // Rota explícita para "/home"
           element={<Home />}
           loader={() => handleVerificationProtected()}
         />
@@ -64,11 +57,6 @@ const router = createBrowserRouter(
       <Route
         path="signup"
         element={<Signup />}
-        loader={() => isAuthenticated()}
-      />
-      <Route
-        path="hometp2"
-        element={<Hometp2 />}
         loader={() => isAuthenticated()}
       />
     </Route>
