+++ src/services/authentication.js
@@ -1,46 +1,14 @@
-import { Navigate, redirect } from "react-router-dom";
+import { redirect } from "react-router-dom";
 
 const isAuthenticated = () => {
-  const session = localStorage.getItem("session");
-
-  if (session) throw redirect("/");
+  const token = localStorage.getItem("token");
+  if (token) throw redirect("/");
   return null;
 };
 
 const handleVerificationProtected = () => {
-  const session = localStorage.getItem("session");
-
-  if (!session) throw redirect("/signin");
+  const token = localStorage.getItem("token");
+  if (!token) throw redirect("/signin");
   return null;
 };
-
-const signIn = async (email, password, supabase) => {
-  return await supabase.auth.signInWithPassword({
-    email,
-    password,
-  });
-};
-
-const signUp = async (email, password, supabase) => {
-  return await supabase.auth.signUp({
-    email,
-    password,
-  });
-};
-
-const signOut = async (supabase, navigate) => {
-  localStorage.removeItem("session");
-  localStorage.removeItem("user");
-
-  supabase.auth.signOut();
-
-  return navigate("/signin");
-};
-
-export {
-  isAuthenticated,
-  handleVerificationProtected,
-  signIn,
-  signUp,
-  signOut,
-};
+export { isAuthenticated, handleVerificationProtected };
