+++ src/views/Forms.tsx
@@ -1,19 +1,5 @@
-import { useParams } from "react-router-dom";
-
 const Forms: React.FC = () => {
-  const params = useParams();
-
-  const getParamType = () => {
-    switch (params.type) {
-      case "1":
-        return "Sono";
-      case "2":
-        return "Amamentação";
-      default:
-        return "Fraldas";
-    }
-  };
-  return getParamType();
+  return "Forms";
 };
 
 export default Forms;
