+++ src/views/Signup.tsx
@@ -1,0 +1,11 @@
+// import { Alert } from "@mui/material";
+import { useAppContext } from "../../src/Context";
+
+const Signup: React.FC = () => {
+  const { showAlertMessage } = useAppContext();
+
+  showAlertMessage("info", "Cadastre-se para obter acesso ao app.");
+  return "Signup";
+};
+
+export default Signup;
